Move memory trapping infrastructure outside of NCE
