+++ app/src/main/cpp/skyline/nce.h
@@ -3,11 +3,8 @@
 
 #pragma once
 
-#include <sys/wait.h>
-#include <linux/elf.h>
 #include "common.h"
 #include "hle/symbol_hooks.h"
-#include "common/interval_map.h"
 
 namespace skyline::nce {
     /**
@@ -18,37 +15,6 @@ namespace skyline::nce {
         const DeviceState &state;
 
         std::vector<hle::HookedSymbol> hookedSymbols; //!< The list of symbols that are hooked, these have a specific ordering that is hardcoded into the hooked functions
-
-        /**
-         * @brief The level of protection that is required for a callback entry
-         */
-        enum class TrapProtection {
-            None = 0, //!< No protection is required
-            WriteOnly = 1, //!< Only write protection is required
-            ReadWrite = 2, //!< Both read and write protection are required
-        };
-
-        using TrapCallback = std::function<bool()>;
-        using LockCallback = std::function<void()>;
-
-        struct CallbackEntry {
-            TrapProtection protection; //!< The least restrictive protection that this callback needs to have
-            LockCallback lockCallback;
-            TrapCallback readCallback, writeCallback;
-
-            CallbackEntry(TrapProtection protection, LockCallback lockCallback, TrapCallback readCallback, TrapCallback writeCallback);
-        };
-
-        std::mutex trapMutex; //!< Synchronizes the accesses to the trap map
-        using TrapMap = IntervalMap<u8*, CallbackEntry>;
-        TrapMap trapMap; //!< A map of all intervals and corresponding callbacks that have been registered
-
-        /**
-         * @brief Reprotects the intervals to the least restrictive protection given the supplied protection
-         */
-        void ReprotectIntervals(const std::vector<TrapMap::Interval>& intervals, TrapProtection protection);
-
-        bool TrapHandler(u8* address, bool write);
 
         static void SvcHandler(u16 svcId, ThreadContext *ctx);
 
@@ -98,8 +64,6 @@ namespace skyline::nce {
          */
         NCE(const DeviceState &state);
 
-        ~NCE();
-
         struct PatchData {
             size_t size; //!< Size of the .patch section
             std::vector<size_t> offsets; //!< Offsets in .text of instructions that need to be patched
@@ -117,41 +81,5 @@ namespace skyline::nce {
         static size_t GetHookSectionSize(span<hle::HookedSymbolEntry> entries);
 
         void WriteHookSection(span<hle::HookedSymbolEntry> entries, span<u32> hookSection);
-
-        /**
-         * @brief An opaque handle to a group of trapped region
-         */
-        class TrapHandle : private TrapMap::GroupHandle {
-            constexpr TrapHandle(const TrapMap::GroupHandle &handle);
-
-            friend NCE;
-        };
-
-        /**
-         * @brief Creates a region of guest memory that can be trapped with a callback for when an access to it has been made
-         * @param lockCallback A callback to lock the resource that is being trapped, it must block until the resource is locked but unlock it prior to returning
-         * @param readCallback A callback for read accesses to the trapped region, it must not block and return a boolean if it would block
-         * @param writeCallback A callback for write accesses to the trapped region, it must not block and return a boolean if it would block
-         * @note The handle **must** be deleted using DeleteTrap before the NCE instance is destroyed
-         * @note It is UB to supply a region of host memory rather than guest memory
-         * @note This doesn't trap the region in itself, any trapping must be done via TrapRegions(...)
-         */
-        TrapHandle CreateTrap(span<span<u8>> regions, const LockCallback& lockCallback, const TrapCallback& readCallback, const TrapCallback& writeCallback);
-
-        /**
-         * @brief Re-traps a region of memory after protections were removed
-         * @param writeOnly If the trap is optimally for write-only accesses, this is not guarenteed
-         */
-        void TrapRegions(TrapHandle handle, bool writeOnly);
-
-        /**
-         * @brief Removes protections from a region of memory
-         */
-        void RemoveTrap(TrapHandle handle);
-
-        /**
-         * @brief Deletes a trap handle and removes the protection from the region
-         */
-        void DeleteTrap(TrapHandle handle);
     };
 }
